requirements.txt
# Core Framework
streamlit>=1.28.0
openai>=1.12.0

# Image Processing
Pillow>=10.0.0
numpy>=1.21.0
opencv-python>=4.5.0

# Video Processing
moviepy>=1.0.3
imageio>=2.19.0
imageio-ffmpeg>=0.4.7
proglog>=0.1.12

# Audio Processing
pydub>=0.25.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
psutil>=5.9.0
tqdm>=4.60.0
PyYAML>=6.0
joblib>=1.1.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For advanced audio processing (uncomment if needed)
# librosa>=0.9.0
# resampy>=0.3.0

# Optional: For AI/ML features (uncomment if needed)
# face-alignment>=1.3.5
# scipy>=1.10.0
# scikit-image>=0.19.0
# kornia>=0.6.0

# Optional: For advanced video processing
# av>=10.0.0  # Uncomment if using advanced video features
# safetensors>=0.3.0  # Uncomment if using model serialization

# Optional: For OCR capabilities (if implementing QA agent OCR)
# pytesseract>=0.3.10  # Uncomment if implementing OCR features
# easyocr>=1.6.0  # Alternative OCR library

# Optional: For face enhancement
# gfpgan>=1.3.0  # Uncomment if using face enhancement
# basicsr>=1.4.0  # Uncomment if using image restoration
# facexlib>=0.3.0  # Uncomment if using face analysis

# Optional: For advanced audio processing
# numba>=0.55.0  # Uncomment if using advanced audio features
# llvmlite>=0.38.0  # Required by numba

# Optional: For configuration management
# yacs>=0.1.8  # Uncomment if using advanced config management

# Note: sqlite3 is built-in with Python, no need to install separately

# Installation Instructions:
# 1. Create a virtual environment: python -m venv venv
# 2. Activate the environment: source venv/bin/activate (Linux/Mac) or venv\Scripts\activate (Windows)
# 3. Install dependencies: pip install -r requirements.txt
# 4. If you encounter issues with moviepy/proglog, try: pip install --force-reinstall moviepy proglog